requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.5.2
astroid==2.11.3
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2019.3.9
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.4
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.4
dj-database-url==0.5.0
Django==4.0.6
django-celery-beat==2.3.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.4
django-redis==5.2.0
django-redis-sessions==0.6.2
django-timezone-field==5.0
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
httptools==0.4.0
idna==3.3
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
matplotlib-inline==0.1.3
mccabe==0.7.0
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pep8==1.7.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.5.0
pycparser==2.21
Pygments==2.12.0
pylint==2.13.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.4
urllib3==1.26.10
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.0
wsproto==1.1.0
